Refactor logging tests to use a unified CountingHandler.
Replaced individual log-specific counters with a single CountingHandler
instance for cleaner and centralized handling of log counts. Simplified
test assertions to validate counts using the new handler's structure.
Minor adjustments were made to improve test clarity and maintainability.
